Update README.md
I have developed a drum kit using HTML, CSS, and JavaScript. The kit includes four tom-toms, a floor tom, a bass drum, and a snare drum, each producing their respective sounds when clicked or when a specific key corresponding to them is pressed. Additionally, the kit features animations to enhance the user experience.